🧪 Added missing tests
